+++ app/Providers/AppServiceProvider.php
@@ -5,6 +5,11 @@ namespace App\Providers;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Support\ServiceProvider;
 
+use TCG\Voyager\Facades\Voyager;
+
+use App\Actions\ApproveSupportTicket;
+use App\Actions\RejectSupportTicket;
+
 class AppServiceProvider extends ServiceProvider
 {
     /**
@@ -29,5 +34,10 @@ class AppServiceProvider extends ServiceProvider
         if (isset($_SERVER['HTTPS']) && ($_SERVER['HTTPS'] == 'on' || $_SERVER['HTTPS'] == 1) || isset($_SERVER['HTTP_X_FORWARDED_PROTO']) &&  $_SERVER['HTTP_X_FORWARDED_PROTO'] == 'https') {
             \URL::forceScheme('https');
         }
+
+        //Reject Support Ticket
+        Voyager::addAction(RejectSupportTicket::class);
+        //Approve Support Ticket
+        Voyager::addAction(ApproveSupportTicket::class);
     }
 }