+++ database/seeders/BreadPermissionsSeeder.php
@@ -3,8 +3,10 @@
 namespace Database\Seeders;
 
 use Illuminate\Database\Seeder;
+
 use TCG\Voyager\Models\Permission;
 use TCG\Voyager\Models\Role;
+use App\Models\PermissionRole;
 
 class BreadPermissionsSeeder extends Seeder
 {
@@ -22,5 +24,20 @@ class BreadPermissionsSeeder extends Seeder
         $role->permissions()->sync(
             $permissions->pluck('id')->all()
         );
+
+        //Remove ability to delete support tickets
+        $roleID = Role::where('name', 'support')->first()->id;
+
+        $permissionIDs[0] = Permission::where('key', 'browse_supporttickets')->first()->id;
+        $permissionIDs[1] = Permission::where('key', 'read_supporttickets')->first()->id;
+
+        foreach ($permissionIDs as $permissionID) {
+            $data = array(
+                'role_id' => $roleID,
+                'permission_id' => $permissionID,
+            );
+
+            PermissionRole::insert($data);
+        }
     }
 }