+++ resources/views/Components/navBar.blade.php
@@ -3,8 +3,8 @@
 <header class="nav-bar">
     <a href="/"><img src="{{ asset('images/logo.png') }}" class="topLogo"></a>
     <div class="nav-bar-center">
-        <a href="/RequestStatus">{!! __('Статус заявки') !!}</a>
-        <a href="/RequestNew" class="red">{!! __('Подать заявку') !!}</a>
+        <a href="/SupportTicketStatus">{!! __('Статус заявки') !!}</a>
+        <a href="/SupportTicketNew" class="red">{!! __('Подать заявку') !!}</a>
     </div>
     <div class="nav-bar-right">
         <img src="{{ asset('images/language.png') }}" class="topLanguage">